feat(UIKIT-877,TreeList): Группировка компонентов в директории Tree
