+++ packages/components/src/Tree/MultipleTreeList/TreeItem/useLogic/utils/checkIsIndeterminate/index.ts
