+++ packages/components/src/Tree/MultipleTreeList/TreeItem/useLogic/utils/checkIsSelected/checkIsSelected.ts
