+++ packages/components/src/Tree/MultipleTreeList/styles.ts
@@ -1,4 +1,4 @@
-import { styled } from '../styles';
+import { styled } from '../../styles';
 
 export const List = styled.ul`
   overflow: hidden;
@@ -7,4 +7,8 @@ export const List = styled.ul`
   padding: 0;
 
   list-style-type: none;
+`;
+
+export const InnerList = styled(List)`
+  overflow: unset;
 `;
