+++ packages/components/src/Tree/TreeList/TreeItem/styles.ts
@@ -1,5 +1,5 @@
-import { Typography } from '../../Typography';
-import { styled } from '../../styles';
+import { Typography } from '../../../Typography';
+import { styled } from '../../../styles';
 
 export const Label = styled(Typography)`
   margin-left: ${({ theme }) => theme.spacing(1)};
