+++ packages/components/src/Tree/TreeList/TreeList.stories.tsx
@@ -1,24 +1,21 @@
 import { useState } from 'react';
 import { type Meta } from '@storybook/react';
 
-import { Typography } from '../Typography';
-import { styled } from '../styles';
+import { Typography } from '../../Typography';
+import { styled } from '../../styles';
 
-import {
-  MultipleTreeList,
-  type MultipleTreeListProps,
-} from './MultipleTreeList';
+import { TreeList, type TreeListProps } from './TreeList';
 
 /**
  * ### [Figma](https://www.figma.com/design/3ghN4WjSgkKx5rETR64jqh/Sirius-Design-System-(%D0%90%D0%9A%D0%A2%D0%A3%D0%90%D0%9B%D0%AC%D0%9D%D0%9E)?node-id=21329-76225&t=ocy6798zA4pD2nYD-0)
  * ### [Guide]()
- * MultipleTreeList - интерактивный компонент, который поддерживает вложенность объектов,
- * позволяет выбрать или множество (multiselect) объектов из списка,
+ * TreeList - интерактивный компонент, который поддерживает вложенность объектов,
+ * позволяет выбрать один объект из списка,
  * свернуть (скрыть) и развернуть (показать) вложенные элементы
  */
-const meta: Meta<typeof MultipleTreeList> = {
-  title: 'Components/MultipleTreeList',
-  component: MultipleTreeList,
+const meta: Meta<typeof TreeList> = {
+  title: 'Components/Tree/TreeList',
+  component: TreeList,
 };
 
 export default meta;
@@ -85,7 +82,7 @@ const FAKE_TREE_LIST_DATA = [
 ];
 
 export const Example = () => {
-  const [value, setValue] = useState<Array<string> | undefined>();
+  const [value, setValue] = useState<string | undefined>();
 
   const fakeData = [
     ...FAKE_TREE_LIST_DATA,
@@ -95,7 +92,7 @@ export const Example = () => {
     },
   ];
 
-  return <MultipleTreeList data={fakeData} value={value} onChange={setValue} />;
+  return <TreeList data={fakeData} value={value} onChange={setValue} />;
 };
 
 const Item = styled.div`
@@ -106,7 +103,7 @@ const Item = styled.div`
 `;
 
 export const RenderItem = () => {
-  const [value, setValue] = useState<Array<string> | undefined>();
+  const [value, setValue] = useState<string | undefined>();
 
   const fakeData = [
     ...FAKE_TREE_LIST_DATA,
@@ -116,7 +113,7 @@ export const RenderItem = () => {
     },
   ];
 
-  const renderItem: MultipleTreeListProps['renderItem'] = (id, label) => (
+  const renderItem: TreeListProps['renderItem'] = ({ id, label }) => (
     <Item>
       <Typography variant="caption">#{id}</Typography>
       <Typography color="textSecondary">{label}</Typography>
@@ -124,7 +121,7 @@ export const RenderItem = () => {
   );
 
   return (
-    <MultipleTreeList
+    <TreeList
       data={fakeData}
       value={value}
       renderItem={renderItem}
