+++ packages/components/src/Tree/index.ts
@@ -1,0 +1,3 @@
+export * from './MultipleTreeList';
+
+export * from './TreeList';
